Refactor database schema and utility functions
